server/pyproject.toml
[tool.poetry]
name = "til"
version = "0.1.0"
description = ""
authors = ["juyeon <ilnyu0756@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
py-ulid = "^1.0.3"
sqlalchemy = "^2.0.36"
mysqlclient = "^2.2.6"
alembic = "^1.14.0"
dependency-injector = "^4.44.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.0"
pydantic = {extras = ["email"], version = "^2.10.5"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
celery = "^5.4.0"
eventlet = "^0.38.2"
pytest-mock = "^3.14.0"
freezegun = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"